Move common test mockup to conftest.py
